+++ NEWS.md
@@ -1,3 +1,14 @@
+# rbibutils 2.2.15 (CRAN)
+
+- adapted some tests to a change in R-devel circa r84986.
+
+
+# rbibutils 2.2.14 (CRAN)
+
+- fixed a couple of tests in 'test-convert.R' which started failing (ca. R svn
+  rev 84760) due to an R-devel change in `person()`.
+
+
 # rbibutils 2.2.13 (CRAN)
 
 - Bugfix: the declaration of the return value of a C function was accidentally